Start splitting parser module up
